+++ CHANGELOG.md
@@ -21,7 +21,7 @@
    - `text-color`
      - [#520](https://github.com/wix-incubator/rich-content/pull/520) color Picker- adding an option to change back to default color
   ### :bug: Bug Fix
-    - `editor`
+   - `editor`
       - [#556](https://github.com/wix-incubator/rich-content/pull/556) fix onAtomicBlockFocus to work when alternating focus between different kinds of plugins
     
 ## 6.6.2 (Dec 11, 2019)