+++ CHANGELOG.md
@@ -12,7 +12,10 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
-  
+
+  ### :bug: Bug Fix
+  - `link`
+    - [#546](https://github.com/wix-incubator/rich-content/pull/546) saves the last data and the initial state of the checkboxes("Open in a new tab", "Add a nofollow tag") is according to the defaults (anchorTarget, relValue)
 </details>
 <hr/>