+++ CHANGELOG.md
@@ -13,13 +13,15 @@
     Click to see more.
   </summary>
 
-  ### :bug: Bug Fix
-  - `editor-common`
-    - [#558](https://github.com/wix-incubator/rich-content/pull/558) getBlockInfo util arguments fixed
   - `link`
     - [#546](https://github.com/wix-incubator/rich-content/pull/546) saves the last data and the initial state of the checkboxes("Open in a new tab", "Add a nofollow tag") is according to the defaults (anchorTarget, relValue)
 </details>
 <hr/>
+
+## 6.6.4 (Dec 30, 2019)
+### :bug: Bug Fix
+- `editor-common`
+  - [#558](https://github.com/wix-incubator/rich-content/pull/558) getBlockInfo util arguments fixed
 
 ## 6.6.3 (Dec 29, 2019)
   ### :rocket: New Feature